requirements.txt
scrapy
pandas
openpyxl
boto3[crt]
amazon-textract-textractor[pdf]
amazon-textract-textractor[pandas]
amazon-textract-textractor[torch]
jsonlines
apache-beam[gcp]
google-cloud-bigquery
google-cloud-storage
pypdf[crypto]
pymupdf
Pillow
send2trash
camelot-py[base]
camelot-py[cv] 
tabula-py
pdfplumber
python-dotenv
